requirements.txt
tqdm==4.36.1
Keras==2.0.8
numpy==1.17.4
matplotlib==3.0.3
scipy==1.3.2
scikit_image==0.15.0
h5py==2.10.0
tensorflow_gpu==1.13.1
imgaug==0.3.0
skimage==0.0
